pyproject.toml
[project]
name = "ratd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "diffusers>=0.33.1",
    "linear-attention-transformer>=0.19.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
]
